[Fix] change file path, css
