+++ vite-project/package.json
@@ -17,16 +17,18 @@
     "@types/react-router-dom": "^5.3.3",
     "@types/react-slider": "^1.3.1",
     "@types/styled-components": "^5.1.26",
+    "axios": "^1.5.0",
     "http-proxy-middleware": "^2.0.6",
     "react": "^18.2.0",
-    "react-datepicker": "^4.16.0",
     "react-cookie": "^6.1.0",
+    "react-datepicker": "^4.16.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.15.0",
     "react-slider": "^2.0.6",
     "styled-components": "^6.0.7"
   },
   "devDependencies": {
+    "@types/axios": "^0.14.0",
     "@types/react": "^18.2.20",
     "@types/react-dom": "^18.2.7",
     "@typescript-eslint/eslint-plugin": "^6.0.0",
