+++ vite-project/src/pages/Show.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const App: React.FC = () => {
+
+    return (
+        <h1>Show component</h1>
+    );
+};
+
+
+export default App;
